pyproject.toml
[project]
name = "lambdaLearn"
license = "Apache-2.0"
version = "1.0.0"
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.16",
    "scipy>=1.5.2",
    "pandas>=1.3.4",
    "scikit-learn>=1.0.2",
    "torch>=1.10.1",
    "torchvision>=0.11.2",
    "torchtext>=0.11.1",
    "matplotlib>=3.5.0",
    "pillow>=8.4.0",
]

[build-system]
requires = [
    "setuptools>=40.8.0", # To support pyproject.toml
    "pip>=21.3", # To support editable build
]
build-backend = "setuptools.build_meta"

[tool.isort]
profile = "black"
